no size_t append in 1.6

+++ db/clientcursor.cpp
@@ -383,9 +383,9 @@ namespace mongo {
 
     void ClientCursor::appendStats( BSONObjBuilder& result ){
         recursive_scoped_lock lock(ccmutex);
-        result.appendNumber("totalOpen", clientCursorsById.size() );
+        result.appendNumber("totalOpen", (int)clientCursorsById.size() );
         result.appendNumber("clientCursors_size", (int) numCursors());
-        result.appendNumber("timedOut" , numberTimedOut);
+        result.appendNumber("timedOut" , (int)numberTimedOut);
     }
     
     // QUESTION: Restrict to the namespace from which this command was issued?
